+++ test_project/many_to_one/models.py
@@ -19,11 +19,12 @@ class Reporter(models.Model):
 class Article(models.Model):
     headline = models.CharField(max_length=100)
     pub_date = models.DateField()
-    reporter = models.ForeignKey(Reporter, related_name='articles',
-                                 on_delete=models.CASCADE)
+    reporter = models.ForeignKey(
+        Reporter, related_name="articles", on_delete=models.CASCADE
+    )
 
     def __str__(self):
         return self.headline
 
     class Meta:
-        ordering = ('headline',)
+        ordering = ("headline",)