+++ tests/filtersets/test_plain.py
@@ -16,16 +16,8 @@ class TestPlainModelFilterSet(object):
                     "restaurant": {
                         "place": 1,
                         "waiters": [
-                            {
-                                "id": 1,
-                                "name": "Joe",
-                                "restaurant": 1
-                            },
-                            {
-                                "id": 2,
-                                "name": "Jonny",
-                                "restaurant": 1
-                            }
+                            {"id": 1, "name": "Joe", "restaurant": 1},
+                            {"id": 2, "name": "Jonny", "restaurant": 1},
                         ],
                         "serves_hot_dogs": True,
                         "serves_pizza": False,
@@ -36,25 +28,21 @@ class TestPlainModelFilterSet(object):
                     "ignored": [{}],
                 }
                 allow_related = False
-                extra_kwargs = {
-                    'id': {'no_lookup': True},
-                }
+                extra_kwargs = {"id": {"no_lookup": True}}
 
         filters = PlaceFilterSet().get_filters()
 
-        assert set(filters.keys()) == {
-            'id', 'name', 'address', 'nicknames',
-        }
+        assert set(filters.keys()) == {"id", "name", "address", "nicknames"}
 
-        assert isinstance(filters['id'], Filter)
-        assert isinstance(filters['id'].form_field, forms.IntegerField)
-        assert filters['id'].no_lookup is True
-        assert isinstance(filters['name'], Filter)
-        assert isinstance(filters['name'].form_field, forms.CharField)
-        assert isinstance(filters['address'], Filter)
-        assert isinstance(filters['address'].form_field, forms.CharField)
-        assert isinstance(filters['nicknames'], Filter)
-        assert isinstance(filters['nicknames'].form_field, forms.CharField)
+        assert isinstance(filters["id"], Filter)
+        assert isinstance(filters["id"].form_field, forms.IntegerField)
+        assert filters["id"].no_lookup is True
+        assert isinstance(filters["name"], Filter)
+        assert isinstance(filters["name"].form_field, forms.CharField)
+        assert isinstance(filters["address"], Filter)
+        assert isinstance(filters["address"].form_field, forms.CharField)
+        assert isinstance(filters["nicknames"], Filter)
+        assert isinstance(filters["nicknames"].form_field, forms.CharField)
 
     def test_get_filters_no_relations_place_diff_source(self):
         class PlaceFilterSet(PlainModelFilterSet):
@@ -64,16 +52,8 @@ class TestPlainModelFilterSet(object):
                     "restaurant": {
                         "place": 1,
                         "waiters": [
-                            {
-                                "id": 1,
-                                "name": "Joe",
-                                "restaurant": 1
-                            },
-                            {
-                                "id": 2,
-                                "name": "Jonny",
-                                "restaurant": 1
-                            }
+                            {"id": 1, "name": "Joe", "restaurant": 1},
+                            {"id": 2, "name": "Jonny", "restaurant": 1},
                         ],
                         "serves_hot_dogs": True,
                         "serves_pizza": False,
@@ -84,27 +64,25 @@ class TestPlainModelFilterSet(object):
                     "ignored": [{}],
                 }
                 allow_related = False
-                fields = ['id', 'name', 'location', 'nicknames']
+                fields = ["id", "name", "location", "nicknames"]
                 extra_kwargs = {
-                    'id': {'no_lookup': True},
-                    'location': {'source': 'address'}
+                    "id": {"no_lookup": True},
+                    "location": {"source": "address"},
                 }
 
         filters = PlaceFilterSet().get_filters()
 
-        assert set(filters.keys()) == {
-            'id', 'name', 'location', 'nicknames',
-        }
+        assert set(filters.keys()) == {"id", "name", "location", "nicknames"}
 
-        assert isinstance(filters['id'], Filter)
-        assert isinstance(filters['id'].form_field, forms.IntegerField)
-        assert filters['id'].no_lookup is True
-        assert isinstance(filters['name'], Filter)
-        assert isinstance(filters['name'].form_field, forms.CharField)
-        assert isinstance(filters['location'], Filter)
-        assert isinstance(filters['location'].form_field, forms.CharField)
-        assert isinstance(filters['nicknames'], Filter)
-        assert isinstance(filters['nicknames'].form_field, forms.CharField)
+        assert isinstance(filters["id"], Filter)
+        assert isinstance(filters["id"].form_field, forms.IntegerField)
+        assert filters["id"].no_lookup is True
+        assert isinstance(filters["name"], Filter)
+        assert isinstance(filters["name"].form_field, forms.CharField)
+        assert isinstance(filters["location"], Filter)
+        assert isinstance(filters["location"].form_field, forms.CharField)
+        assert isinstance(filters["nicknames"], Filter)
+        assert isinstance(filters["nicknames"].form_field, forms.CharField)
 
     def test_get_filters_with_both_reverse_and_direct_relations(self):
         class PlaceFilterSet(PlainModelFilterSet):
@@ -114,16 +92,8 @@ class TestPlainModelFilterSet(object):
                     "restaurant": {
                         "place": 1,
                         "waiters": [
-                            {
-                                "id": 1,
-                                "name": "Joe",
-                                "restaurant": 1
-                            },
-                            {
-                                "id": 2,
-                                "name": "Jonny",
-                                "restaurant": 1
-                            },
+                            {"id": 1, "name": "Joe", "restaurant": 1},
+                            {"id": 2, "name": "Jonny", "restaurant": 1},
                         ],
                         "serves_hot_dogs": True,
                         "serves_pizza": False,
@@ -132,33 +102,38 @@ class TestPlainModelFilterSet(object):
                     },
                     "nicknames": ["dogs"],
                     "name": "Demon Dogs",
-                    "address": "944 W. Fullerton"
+                    "address": "944 W. Fullerton",
                 }
-                extra_kwargs = {
-                    'restaurant': {
-                        'place': {'no_lookup': True},
-                    },
-                }
+                extra_kwargs = {"restaurant": {"place": {"no_lookup": True}}}
 
         filters = PlaceFilterSet().get_filters()
 
         assert set(filters.keys()) == {
-            'id', 'address', 'name', 'nicknames', 'restaurant',
+            "id",
+            "address",
+            "name",
+            "nicknames",
+            "restaurant",
         }
-        assert set(filters['restaurant'].filters.keys()) == {
-            'place', 'serves_hot_dogs', 'serves_pizza', 'waiters',
+        assert set(filters["restaurant"].filters.keys()) == {
+            "place",
+            "serves_hot_dogs",
+            "serves_pizza",
+            "waiters",
         }
-        assert set(filters['restaurant'].filters['waiters'].filters.keys()) == {
-            'id', 'name', 'restaurant',
+        assert set(filters["restaurant"].filters["waiters"].filters.keys()) == {
+            "id",
+            "name",
+            "restaurant",
         }
 
-        assert filters['restaurant'].filters['place'].no_lookup is True
+        assert filters["restaurant"].filters["place"].no_lookup is True
 
-        assert isinstance(filters['id'], Filter)
-        assert isinstance(filters['id'].form_field, forms.IntegerField)
-        assert isinstance(filters['address'], Filter)
-        assert isinstance(filters['address'].form_field, forms.CharField)
-        assert isinstance(filters['name'], Filter)
-        assert isinstance(filters['name'].form_field, forms.CharField)
-        assert isinstance(filters['restaurant'], PlainModelFilterSet)
-        assert isinstance(filters['restaurant'].filters['waiters'], PlainModelFilterSet)
+        assert isinstance(filters["id"], Filter)
+        assert isinstance(filters["id"].form_field, forms.IntegerField)
+        assert isinstance(filters["address"], Filter)
+        assert isinstance(filters["address"].form_field, forms.CharField)
+        assert isinstance(filters["name"], Filter)
+        assert isinstance(filters["name"].form_field, forms.CharField)
+        assert isinstance(filters["restaurant"], PlainModelFilterSet)
+        assert isinstance(filters["restaurant"].filters["waiters"], PlainModelFilterSet)